# specs/002_api_specification.yaml
...
openapi: "3.1.0"
info:
  title: Cyber Chakra Forensics Admin Portal API
  version: "1.0.0"
  description: |
    REST API for the CMF Admin Portal and the public-facing desktop app endpoints.
    All admin endpoints require Bearer token authentication.
    All public endpoints require a valid license key (X-License-Key header).
servers:
  - url: https://cyberchakra.online
    description: Production
  - url: https://staging.cyberchakra.online
    description: Staging

# =============================================================================
# Security Schemes
# =============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin portal JWT token obtained from /api/v1/auth/login.

    LicenseKeyAuth:
      type: apiKey
      in: header
      name: X-License-Key
      description: License key for public desktop app endpoints.

  # ===========================================================================
  # Shared Schemas
  # ===========================================================================
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 25
        total:
          type: integer
          example: 142
        total_pages:
          type: integer
          example: 6
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
    Release:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: string
        version_code:
          type: integer
        channel:
          type: string
          enum: [stable, beta, alpha, nightly, lts]
        status:
          type: string
          enum: [draft, published, deprecated, yanked]
        title:
          type: string
        release_notes:
          type: string
        is_mandatory:
          type: boolean
        is_security:
          type: boolean
        published_at:
          type: string
          format: date-time
        assets:
          type: array
          items:
            $ref: '#/components/schemas/ReleaseAsset'
        created_at:
          type: string
          format: date-time
    ReleaseAsset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        release_id:
          type: string
          format: uuid
        platform:
          type: string
          enum: [windows_x64, windows_arm64, linux_x64, linux_arm64, macos_x64, macos_arm64]
        filename:
          type: string
        file_size:
          type: integer
        content_type:
          type: string
        sha256_hash:
          type: string
        installer_type:
          type: string
        created_at:
          type: string
          format: date-time

    # ---- Announcement ----
...
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QueryPage:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    QueryPerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 25
        minimum: 1
        maximum: 100
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/releases:
    get:
      tags: [Releases]
      summary: List releases
      parameters:
        - $ref: '#/components/parameters/QueryPage'
        - $ref: '#/components/parameters/QueryPerPage'
        - name: channel
          in: query
          schema:
            type: string
            enum: [stable, beta, alpha, nightly, lts]
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, deprecated, yanked]
      responses:
        "200":
          description: Releases list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Release'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
    post:
      tags: [Releases]
      summary: Create release (draft)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version, version_code, title]
              properties:
                version:
                  type: string
                  description: "Semver string, e.g. 2.1.0"
                version_code:
                  type: integer
                  description: "Monotonic integer for comparison"
                channel:
                  type: string
                  enum: [stable, beta, alpha, nightly, lts]
                  default: stable
                title:
                  type: string
                release_notes:
                  type: string
                  description: "Markdown"
                min_os_version:
                  type: object
                min_upgrade_from:
                  type: string
                is_mandatory:
                  type: boolean
                  default: false
                is_security:
                  type: boolean
                  default: false
      responses:
        "201":
          description: Release created in draft status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        "409":
          description: Version already exists
  /api/v1/releases/{id}:
    get:
      tags: [Releases]
      summary: Get release details
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        "200":
          description: Release with assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
    patch:
      tags: [Releases]
      summary: Update release (draft only)
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                release_notes:
                  type: string
                channel:
                  type: string
                is_mandatory:
                  type: boolean
                is_security:
                  type: boolean
                min_os_version:
                  type: object
                min_upgrade_from:
                  type: string
      responses:
        "200":
          description: Release updated
        "409":
          description: Cannot update published release
  /api/v1/releases/{id}/publish:
    post:
      tags: [Releases]
      summary: Publish a release
      description: |
        Transitions a draft release to published status.
        Requires at least one asset to be uploaded.
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        "200":
          description: Release published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        "409":
          description: Release has no assets or is already published
...